Implement package fetch retries
There are some locations where memory is being freed that wasn't being
freed before. This is because, in the case of the user possibly
configuring `Fetch` to retry many times, I don't want the allocator to
allocate without freeing each time it attempts to refetch the network
data, resulting in a memory leak.