crypto.ghash_polyval: use the R/F algorithm
A new algorithm for efficient GHASH and POLYVAL computation was
just published: https://eprint.iacr.org/2025/2171.pdf

It provides a very nice speedup on AARCH64 (more than 2x) and
a ~10% speedup on Intel.

The downside is a slight increase of memory usage and more
precomputation, but this is generally worth it.