std.time.epoch: use u4 instead of enum for getDaysInMonth
The only way to use `getDaysInMonth` with an integer month is to first
cast to the Month enum. We should switch the default to accept an
integer instead of an enum because the enum already has a built-in
`.numeric()` function. It's much simpler in application code to convert
a Month enum to an integer than vice versa.