`std.math`: add vector support to `wrap`
Breaking because the function changed from `wrap(x: anytype, r: anytype)` to `wrap(x; anytype, r: @TypeOf(x))`,
though looking at examples on github, this doesn't seem to be an issue.
It is likely also possible to implement this with the current signature, but handling two types and figuring out
which of the two to use for which check would be complex.
The main usecases of `x` and `r` being different types is when wrapping a float around an integer, though from
what I've seen people only do that for `comptime_integer`s, so should be fine, and having `x` be a signed integer
and `r` be an unsigned one with the same bitwidth and a value above the maximu value of `@TypeOf(x)`, though
I don't think the function would really do anything then anyway.